Second patch (fix of the previous)
So apperantly I cannot change names of files in git without making some big changes in the filesystem so I renamed the files as they were - with capital letters. In the meantime i made a pull request trying to rename a file directly from github, but that didnt work as well, so I will left them like this.